```
benchmark           old ns/op     new ns/op     delta
BenchmarkIncr-8     5955          3808          -36.05%

benchmark           old allocs     new allocs     delta
BenchmarkIncr-8     18             8              -55.56%

benchmark           old bytes     new bytes     delta
BenchmarkIncr-8     976           577           -40.88%
```